pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cosmic-cupid"
version = "0.1.0"
description = "An AI-powered cosmic Valentine card generator for the terminal."
dependencies = [
    "rich",
    "kerykeion",
    "Pillow"
]

[project.scripts]
cosmic-cupid = "src.main:main"